整理

+++ YanXiuStudentApp/YanXiuStudentApp/v1.2/Views/YXMistakeContentChapterKnpViewController.m
@@ -76,7 +76,7 @@
     }];
 }
 
-- (void)requestForErrorsList:(int)page {
+- (void)requestForErrorsList:(NSInteger)page {
     if (self.isLoading) {
         return;
     }
